+++ core-ng/src/main/java/core/framework/api/log/Markers.java
@@ -1,13 +1,13 @@
 package core.framework.api.log;
 
-import core.framework.impl.log.marker.ErrorTypeMarker;
+import core.framework.impl.log.marker.ErrorCodeMarker;
 import org.slf4j.Marker;
 
 /**
  * @author neo
  */
 public final class Markers {
-    public static Marker errorType(String errorType) {
-        return new ErrorTypeMarker(errorType);
+    public static Marker errorCode(String code) {
+        return new ErrorCodeMarker(code);
     }
 }