+++ core-ng/src/main/java/core/framework/impl/log/marker/ErrorCodeMarker.java
@@ -3,17 +3,17 @@ package core.framework.impl.log.marker;
 /**
  * @author neo
  */
-public final class ErrorTypeMarker extends AbstractMarker {
+public final class ErrorCodeMarker extends AbstractMarker {
     private static final long serialVersionUID = 2633013025655449295L;
 
-    private final String type;
+    private final String code;
 
-    public ErrorTypeMarker(String type) {
-        this.type = type;
+    public ErrorCodeMarker(String code) {
+        this.code = code;
     }
 
     @Override
     public String getName() {
-        return type;
+        return code;
     }
 }