+++ core-ng/src/main/java/core/framework/impl/log/queue/ActionLogMessage.java
@@ -25,8 +25,8 @@ public class ActionLogMessage {
     public String refId;
     @XmlElement(name = "action")
     public String action;
-    @XmlElement(name = "error_type")
-    public String errorType;
+    @XmlElement(name = "error_code")
+    public String errorCode;
     @XmlElement(name = "error_message")
     public String errorMessage;
     @XmlElement(name = "elapsed")