+++ core-ng/src/main/java/core/framework/impl/mongo/MongoImpl.java
@@ -268,13 +268,13 @@ public final class MongoImpl implements Mongo, MongoOption {
 
     private void checkSlowQuery(long elapsedTime) {
         if (elapsedTime > slowQueryThresholdInMs) {
-            logger.warn(Markers.errorType("SLOW_QUERY"), "slow mongo query, elapsedTime={}", elapsedTime);
+            logger.warn(Markers.errorCode("SLOW_QUERY"), "slow mongo query, elapsedTime={}", elapsedTime);
         }
     }
 
     private void checkTooManyRowsReturned(int size) {
         if (size > tooManyRowsReturnedThreshold) {
-            logger.warn(Markers.errorType("TOO_MANY_ROWS_RETURNED"), "too many rows returned, returnedRows={}", size);
+            logger.warn(Markers.errorCode("TOO_MANY_ROWS_RETURNED"), "too many rows returned, returnedRows={}", size);
         }
     }