+++ core-ng/src/main/java/core/framework/impl/redis/RedisImpl.java
@@ -313,7 +313,7 @@ public final class RedisImpl implements Redis {
 
     private void checkSlowQuery(long elapsedTime) {
         if (elapsedTime > slowQueryThresholdInMs) {
-            logger.warn(Markers.errorType("SLOW_QUERY"), "slow redis query, elapsedTime={}", elapsedTime);
+            logger.warn(Markers.errorCode("SLOW_QUERY"), "slow redis query, elapsedTime={}", elapsedTime);
         }
     }
 }